Code for numbers divisible by 7
Have written code to check numbers that are divisible by 7 and check to make sure it passes. Had an error thrown up as the number I used in the test was divisible by 3 and 7, so have had to make the necessary adjustments. Also, have had to change my code in place in my two files with my code in them, as I noticed last time I was committing that my tests were in the wrong file. Moving on to writenecessary tests for numbers that are divisible by either two of the numbers or all 3 numbers.